+++ libft/libft.h
@@ -22,7 +22,7 @@ typedef struct s_list
 	struct s_list	*next;
 }				t_list;
 
-int			ft_atoi(const char *str);
+long long	ft_atoi(const char *str);
 
 void		ft_bzero(void *s, size_t n);
